move debugger and coverage to extentions
