+++ qiling/core.py
@@ -16,7 +16,7 @@ from .utils import ql_setup_logging_env, ql_setup_logging_stream
 from .core_struct import QLCoreStructs
 from .core_hooks import QLCoreHooks
 from .core_utils import QLCoreUtils
-from .debugger import ql_debugger_init
+from .extensions.debugger import ql_debugger_init
 
 __version__ = "1.1" + "-dev"
