+++ qiling/extensions/debugger/gdbserver/xml/x86/32bit-avx512.xml
