+++ qiling/extensions/debugger/gdbserver/xml/x86/32bit-segments.xml
